requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.13
faiss-cpu==1.7.4
streamlit==1.31.0
python-dotenv==1.0.0
pypdf==3.17.1
sentence-transformers==2.2.2
chromadb==0.4.22
tiktoken==0.5.2
pytest==7.4.3
black==23.12.1
isort==5.13.2
flake8==7.0.0 